% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_procrustes.R
\name{apply_procrustes}
\alias{apply_procrustes}
\title{Procrustes alignment and mapping back to distances}
\usage{
apply_procrustes(X_base, Y_base, Y)
}
\arguments{
\item{X_base}{Base coordinates for target alignment}

\item{Y_base}{Base coordinates for source alignment}

\item{Y}{Full source coordinates to transform}
}
\value{
Transformed coordinates matrix
}
\description{
Procrustes alignment and mapping back to distances
}
