% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besmi.R
\name{besmi_knn_impute}
\alias{besmi_knn_impute}
\title{KNN imputation sweep (uses VIM::kNN)}
\usage{
besmi_knn_impute(
  M_input,
  M_mask,
  M_real = NULL,
  distance_metric = "mae",
  k = NA,
  bs_i = NA
)
}
\arguments{
\item{M_input}{Matrix with NAs}

\item{M_mask}{Logical mask matrix}

\item{M_real}{Optional ground truth}

\item{distance_metric}{Distance metric name}

\item{k}{Dataset parameter k}

\item{bs_i}{Bootstrap index}
}
\value{
List with final_matrix, metrics, tails_chain
}
\description{
KNN imputation sweep (uses VIM::kNN)
}
