% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besmi.R
\name{besmi_prepare_full_dataset}
\alias{besmi_prepare_full_dataset}
\title{Prepare full GDM dataset from CSV or RData}
\usage{
besmi_prepare_full_dataset(input_path)
}
\arguments{
\item{input_path}{Path to CSV or RData file containing the full distance matrix}
}
\value{
Symmetric numeric matrix
}
\description{
Prepare full GDM dataset from CSV or RData
}
