% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_procrustes.R
\name{coords_to_distances}
\alias{coords_to_distances}
\title{Convert coordinate matrix to distance matrix}
\usage{
coords_to_distances(coords)
}
\arguments{
\item{coords}{Numeric coordinate matrix}
}
\value{
Symmetric distance matrix
}
\description{
Convert coordinate matrix to distance matrix
}
