% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gdm.R
\name{create_mds_plot}
\alias{create_mds_plot}
\title{Create MDS plot of genetic distances}
\usage{
create_mds_plot(dist_matrix, pop_info)
}
\arguments{
\item{dist_matrix}{Symmetric numeric distance matrix}

\item{pop_info}{Data frame with metadata columns}
}
\value{
A ggplot object
}
\description{
Create MDS plot of genetic distances
}
