% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besmi.R
\name{.besmi_determine_sampling_sizes}
\alias{.besmi_determine_sampling_sizes}
\title{Determine bootstrap sample count for a given k}
\usage{
.besmi_determine_sampling_sizes(k)
}
\description{
Determine bootstrap sample count for a given k
}
\keyword{internal}
