% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_procrustes.R
\name{perform_mds}
\alias{perform_mds}
\title{Perform MDS on a pair of distance matrices}
\usage{
perform_mds(A, B)
}
\arguments{
\item{A}{First distance matrix}

\item{B}{Second distance matrix}
}
\value{
A list with coordinates \code{X}, \code{Y}, optimal dimension \code{d_opt}, and variance info
}
\description{
Perform MDS on a pair of distance matrices
}
