% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gdm.R
\name{visualize_results}
\alias{visualize_results}
\title{Create plotting handles for simulation results}
\usage{
visualize_results(sim_results)
}
\arguments{
\item{sim_results}{A list returned by simulate_genetic_distances() or run_genetic_simulation()}
}
\value{
A list with \code{heatmap} and \code{mds} functions that print plots when called
}
\description{
Create plotting handles for simulation results
}
